Own_Agent_v2.0/requirements.txt
fastapi
uvicorn
crewai
crewai-tools
python-dotenv
pydantic
mcp
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
xgboost
shap
crewai[tools]
langchain
langchain-community
jupyter
ipykernels
httpx
requests
loguru
chromadb
faiss-cpu